test: compare a c_char without a cast that is wrong on half the targets
`buf.iter().all(|&c| c as u8 == 0x7F)` in tests/capi.rs is necessary on
x86_64, where `c_char` is signed, and a clippy error on aarch64, where it
is unsigned ("casting to the same type is unnecessary"). So the
repository could not be committed to from an arm64 host with the
pre-commit guards installed, and the lint is real on that host rather
than a false positive.

`i32::from(c)` widens from either sign, needs no cast, and compares the
same value: a byte the readlink refusal must have left at 0x7F.

+++ tests/capi.rs
@@ -316,7 +316,11 @@ fn readlink_refuses_a_buffer_too_small_for_the_target() {
         last_errno_erange()
     );
     assert!(
-        buf.iter().all(|&c| c as u8 == 0x7F),
+        // `i32::from`, not `c as u8`: `c_char` is SIGNED on x86_64 and
+        // UNSIGNED on aarch64, so a cast that is necessary on one is a
+        // clippy error on the other and the pre-commit guard cannot pass
+        // on both. Widening is defined for either sign and needs no cast.
+        buf.iter().all(|&c| i32::from(c) == 0x7F),
         "a refused readlink must not have written into the buffer"
     );
     unsafe { fs_xfs_umount(fs) };
